feat: split novelildocument class to new file for improving
